requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.22.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
Django==5.1
djangorestframework==3.15.2
idna==3.7
isodate==0.6.1
msal==1.30.0
msal-extensions==1.2.0
pillow==10.4.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
srt==3.5.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vosk==0.3.45
websockets==12.0
